✨ feat - Refreshes the initialize pages.
