+++ Source/Mobile/Finance/AppShell.xaml
@@ -2,10 +2,11 @@
 <Shell xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
        xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
        xmlns:pages="clr-namespace:Finance.Pages"
+       xmlns:pages_initialize="clr-namespace:Finance.Pages.Initialize"
        x:Class="Finance.AppShell">
 
     <!-- Page Registration -->
-    <ShellContent FlyoutItemIsVisible="False" Route="loading" ContentTemplate="{DataTemplate pages:LoadingPage}" />
-    <ShellContent FlyoutItemIsVisible="False" Route="presentation" ContentTemplate="{DataTemplate pages:WalletPresentationPage}" />
+    <ShellContent FlyoutItemIsVisible="False" Route="loading" ContentTemplate="{DataTemplate pages_initialize:LoadingPage}" />
+    <ShellContent FlyoutItemIsVisible="False" Route="presentation" ContentTemplate="{DataTemplate pages_initialize:PresentationPage}" />
     <ShellContent FlyoutItemIsVisible="False" Route="main" ContentTemplate="{DataTemplate pages:MainPage}" />
 </Shell>
