+++ Source/Mobile/Finance/AppShell.xaml.cs
@@ -1,6 +1,4 @@
 ﻿using Finance.Pages.DangerZone;
-using Finance.Pages.Historic;
-using Finance.Pages.Strategy;
 using Finance.Pages.Walleting;
 
 namespace Finance;
@@ -12,9 +10,6 @@ public partial class AppShell : Shell {
 
         Routing.RegisterRoute("create", typeof(CreateWalletPage));
         Routing.RegisterRoute("select", typeof(SelectWalletPage));
-
-        Routing.RegisterRoute("strategy", typeof(StrategyPage));
-        Routing.RegisterRoute("historic", typeof(HistoricPage));
         Routing.RegisterRoute("dangerZone", typeof(DangerZonePage));
         Routing.RegisterRoute("dangerZone/delete", typeof(DeleteWalletPage));
     }
