--- Source/Mobile/Finance/Pages/Historic/HistoricPage.xaml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:buttons="clr-namespace:Finance.Controls.Buttons"
-             xmlns:navigation="clr-namespace:Finance.Controls.Navigation"
-             x:Class="Finance.Pages.Historic.HistoricPage"
-             Shell.NavBarIsVisible="False"
-             Shell.FlyoutBehavior="Disabled">
-
-    <Grid RowDefinitions="Auto,Auto,*">
-        <navigation:NavigationBar Title="Histórico" IsRoot="False" Clicked="BackButtom_Clicked">
-            <navigation:NavigationBar.ToolbarItems>
-                <VerticalStackLayout>
-                    <navigation:ToolbarItem Glyph="&#x2b;" Command="{Binding NewOperationCommand}" />
-                </VerticalStackLayout>
-            </navigation:NavigationBar.ToolbarItems>
-        </navigation:NavigationBar>
-        <!--<Border Grid.Row="2" Margin="15,20" Style="{StaticResource Card}">
-            <editors:DXCalendar BackgroundColor="{AppThemeBinding Light={StaticResource Gray200}, Dark={StaticResource Gray600}}"
-                                MinDate="{Binding MinDate}" MaxDate="{Binding MaxDate}" DisplayDate="{Binding DisplayDate}" SelectedDate="{Binding SelectedDate}">
-                <editors:DXCalendar.HeaderAppearance>
-                    <editors:CalendarHeaderAppearance HeaderTitleTextColor="{AppThemeBinding Light={StaticResource Gray600}, Dark={StaticResource Gray000}}"
-                                                      HeaderSubtitleTextColor="{AppThemeBinding Light={StaticResource Gray600}, Dark={StaticResource Gray000}}"
-                                                      BackgroundColor="{AppThemeBinding Light={StaticResource Gray300}, Dark={StaticResource Gray600}}"
-                                                      HeaderTitleFontAttributes="None" />
-                </editors:DXCalendar.HeaderAppearance>
-                <editors:DXCalendar.DayCellAppearance>
-                    <editors:CalendarDayCellAppearance TextColor="{AppThemeBinding Light={StaticResource Gray600}, Dark={StaticResource Gray000}}"
-                                                       DisabledTextColor="{AppThemeBinding Light={StaticResource Gray400}, Dark={StaticResource Gray500}}"
-                                                       SelectedFontAttributes="Bold"
-                                                       TodayEllipseBackgroundColor="Transparent"
-                                                       TodayTextColor="{AppThemeBinding Light={StaticResource Gray600}, Dark={StaticResource Gray000}}" />
-                </editors:DXCalendar.DayCellAppearance>
-            </editors:DXCalendar>
-        </Border>-->
-    </Grid>
-</ContentPage>
