+++ Source/Mobile/Finance/Pages/Initialize/LoadingPage.xaml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Finance.Pages.LoadingPage"
+             x:Class="Finance.Pages.Initialize.LoadingPage"
              Shell.FlyoutBehavior="Disabled">
 </ContentPage>
