--- Source/Mobile/Finance/Pages/Strategy/StrategyPage.xaml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:buttons="clr-namespace:Finance.Controls.Buttons"
-             xmlns:navigation="clr-namespace:Finance.Controls.Navigation"
-             x:Class="Finance.Pages.Strategy.StrategyPage"
-             Shell.NavBarIsVisible="False"
-             Shell.FlyoutBehavior="Disabled">
-
-    <Grid RowDefinitions="Auto,*">
-        <navigation:NavigationBar Title="Estratégia" IsRoot="False" Clicked="BackButtom_Clicked" />
-        <ScrollView Grid.Row="1" VerticalScrollBarVisibility="Never">
-            <VerticalStackLayout Spacing="20" Margin="15,0,15,50">
-                <Border Style="{x:StaticResource Card}">
-                    <Grid ColumnDefinitions="*,Auto">
-                        <Label Text="Ações Brasil" VerticalTextAlignment="Center" Margin="10,0" FontSize="Medium" FontAttributes="Bold" />
-                        <buttons:IndicatorButton Grid.Column="1" Text="Desativar" WidthRequest="100" IsRunning="{Binding IsBrazilStocksRunning}" Command="{Binding BrazilStocksEnabledCommand}" />
-                    </Grid>
-                </Border>
-                <Border Style="{x:StaticResource Card}">
-                    <Grid ColumnDefinitions="*,Auto">
-                        <Label Text="FII" VerticalTextAlignment="Center" Margin="10,0" FontSize="Medium" FontAttributes="Bold" />
-                        <buttons:IndicatorButton Grid.Column="1" Text="Desativar" WidthRequest="100" IsRunning="{Binding IsFIIsRunning}" Command="{Binding FIIsEnabledCommand}" />
-                    </Grid>
-                </Border>
-            </VerticalStackLayout>
-        </ScrollView>
-    </Grid>
-</ContentPage>
