+++ Source/Mobile/Finance/Pages/Walleting/CreateWalletViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using Finance.Models;
+using Finance.Pages.Initialize;
 using Finance.Validators.Objects;
 using Finance.Validators.Rules;
