+++ Source/Mobile/Finance/Pages/Walleting/SelectWalletViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using Finance.Models;
+using Finance.Pages.Initialize;
 
 namespace Finance.Pages.Walleting;
